feature: application optimization (#20)
* feature(host): add default host as 0.0.0.0 for flask app

* feature(dockerignore): add files and folder docker will ignore

* feature(docker-compose.yml): set compose configurations for client

* feature(docker-vue): set dockerFile configs for vue app

* feature(docker-flask): set dockerFile configs for flask app

* feature(vue-env): replace backend url with VUE_APP_BACKEND_URL variable that get from env

* refactor: Remove salt field from User model and update docstring

* refactor: Update comments in authentication module and remove salt fields of user

* refactor: Add missing import prefix

* refactor: remove unused resource

* refactor: Add missing import prefix and handle new errors of mongoengine

* refactor: Add missing import prefix and make corrections on error messages

+++ server/app/resources/post.py
@@ -1,5 +1,6 @@
 from flask_restful import Resource, reqparse
-from models.post import Post
+from app.models.post import Post
+from mongoengine.errors import ValidationError, DoesNotExist
 from flask_jwt_extended import current_user, jwt_required
 import ast
 
@@ -114,14 +115,18 @@ class PostResource(Resource):
         """
 
         try:
+            if not Post.objects.get(id=post_id):
+                return {"error": "Post not found"}, 403
             post = Post.objects.get(id=post_id)
-
             if post.author.id == current_user.id:
                 post.delete()
                 return {"message": "Post deleted successfully"}
             else:
                 return {"error": "You don't have permission to delete this post"}, 403
-        except Post.DoesNotExist:
-            return {"error": "Post not found"}, 404
+        except DoesNotExist:
+            return {"error": "Post not found"}, 403
+        except ValidationError:
+            return {"error": "Post not found"}, 403    
         except Exception as e:
+            print(e)
             return {"error": f"An unexpected error occurred: {str(e)}"}, 500
